LEGEND: Added Gateway descendant Parser class
I identified a bunch of constants the parser was using
as common words like all, and, again, etc. Since each
game has their own vocab list, it's more than likely
that the Ids for these words differ between games.

Because of this, I set up the main parser class to have
a bunch of constants that the descendant parser for
Gateway fills out with the correct ids. Other games
implemented later will be able to do likewise