WASTELAND: Refactoring script handler, adding doneCallback
The original script interpreter has opcodes that may do
things like prompt the user for Y/N actions when changing
maps. As such, I'll need to set up script handling so
it can pause execution, show a different view, and then
the view can pass results back to the script and have it
resume. Then return to the original calling code that
handles the party movement