WASTELAND: Fix reading script offsets
Initially I set up srcipt offsets as a separate table, assuming
that the first offset would point to just after the end of the
table, so I could use it to get the table size. However, the
overworld map has 0 entries at the start of the block,
so there's no point in maintaining offsets as a separate table.