DINK: In progress changing over to 32-bit game surface
I initially had it as a paletted surface, however things were wonky.
The colors 0 and 255 are reserved for Black and white (I think)
by DirectDraw. However, the palette of the splash screen has
index 0 as white instead of black. If I hardcoded it to black,
then the splash screen didn't display correctly. But if I leave it
as white, then it paints the background of the following main menu
as white instead of black.

I'm thus switching the engine to high color to avoid these issues.