SPYCRAFT: Added duck_api.cpp
This is just a stub for several functions at the moment.
The original source headers mention the following:
sync objects in both EXE and DLL ugly -- will change for sure

I can only assume that some of the sound code was done as a
separate DLL, maybe interfacing to a third party Duck sound lib.